[config] Split `runtime` and `development`
